Added ui config, decoupled tests
