+++ src/adapters/api/controllers/routes.rs
@@ -1,10 +1,10 @@
 use axum::middleware::from_fn;
 use utoipa_axum::router::OpenApiRouter;
 use utoipa_axum::routes;
+use crate::commands::startup::get_api_config;
 use crate::constants::inner_constants::ENVIRONMENT_SERVICE;
 use crate::gpodder::auth::authentication::login;
 use crate::gpodder::parametrization::{get_client_parametrization_router};
-use crate::{get_api_config};
 use crate::gpodder::device::device_controller::get_device_router;
 use crate::gpodder::episodes::gpodder_episodes::get_gpodder_episodes_router;
 use crate::gpodder::session_middleware::handle_cookie_session;
