+++ src/controllers/podcast_episode_controller.rs
@@ -15,10 +15,9 @@ use crate::models::podcast_dto::PodcastDto;
 use crate::models::settings::Setting;
 use crate::service::file_service::perform_episode_variable_replacement;
 use std::thread;
-use axum::{debug_handler, Extension, Json};
+use axum::{Extension, Json};
 use axum::extract::{Path, Query};
 use axum::http::StatusCode;
-use axum::response::IntoResponse;
 use utoipa_axum::router::OpenApiRouter;
 use utoipa_axum::routes;
 use crate::controllers::server::ChatServerHandle;
